backend/requirements.txt
Django==5.1.2
django-allauth==65.1.0
django-environ==0.11.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
jwcrypto==1.5.6
psycopg==3.2.3
python-keycloak==4.7.0
django-cors-headers==4.6.0
gunicorn==23.0.0